addTour() and removeTour() for AdminController
