+++ VladShalamai/src/travelagency/tests/TestAdminController.java
@@ -1,5 +1,6 @@
 package travelagency.tests;
 
+import travelagency.controller.AdminController;
 import travelagency.db.DataBase;
 import travelagency.model.*;
 
@@ -9,11 +10,12 @@ import static utils.TravelAgencyUtils.*;
 /**
  * Created by Влад on 28.11.2016.
  */
-public class TestDataBase {
+public class TestAdminController {
 
     public static void main(String[] args) {
 
         DataBase dataBase = new DataBase();
+
         for (int i = 0; i < 5; i++) {
             dataBase.getTours().add(generateTour());
         }
@@ -25,8 +27,10 @@ public class TestDataBase {
 
     private static void testAddTour(DataBase dataBase) {
 
-        dataBase.addTour("Relax in Paris", 3000, new MyDate(2016, 3, 2, new MyTime(12, 10)),
-                        new MyDate(2016, 3, 20, new MyTime(12, 10)), "plane", new Hotel("Hilton",
+        AdminController admin = new AdminController(dataBase);
+
+        admin.addTour("Relax in Paris", 3000, new MyDate(2016, 3, 2, new MyTime(12, 10)),
+                new MyDate(2016, 3, 20, new MyTime(12, 10)), "plane", new Hotel("Hilton",
                         new Address("USA", "New-York", "Wall-Street", 12), 700, 5));
 
         System.out.println("addTour() is " + (dataBase.getTours().size() == 6));
@@ -35,7 +39,9 @@ public class TestDataBase {
 
     private static void testRemoveTour(DataBase dataBase) {
 
-        Tour removed = dataBase.removeTour(2);
+        AdminController admin = new AdminController(dataBase);
+
+        Tour removed = admin.removeTour(2);
         System.out.printf("removeTour() is " + (removed.getId() == 2));
     }
